requirements.txt
numpy
tensorflow
scipy
opencv-python
keras
Pillow